# gui/src/assets/openapi.json
{
    "openapi": "3.1.0",
    "info": {
    "title": "FastAPI",
    "version": "0.1.0"
    },
    "paths": {
    "/api/v1/auth/forgot-password": {
    "post": {
    "tags": [
    "auth"
    ],
    "summary": "Authcontroller.Forgot Password",
    "operationId": "AuthController_forgot_password_api_v1_auth_forgot_password_post",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "type": "string",
    "title": "Username"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/auth/login": {
    "post": {
    "tags": [
    "auth"
    ],
    "summary": "Authcontroller.Login",
    "operationId": "AuthController_login_api_v1_auth_login_post",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/LoginCredentials"
    }
    }
    },
    "required": true
    },
    "responses": {
    "204": {
    "description": "Successful response"
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/auth/logout": {
    "post": {
    "tags": [
    "auth"
    ],
    "summary": "Authcontroller.Logout",
    "operationId": "AuthController_logout_api_v1_auth_logout_post",
    "responses": {
    "204": {
    "description": "Successful response"
    }
    },
    "security": [
    {
    "APIKeyCookie": []
    }
    ]
    }
    },
    "/api/v1/auth/me": {
    "get": {
    "tags": [
    "auth"
    ],
    "summary": "Authcontroller.Me",
    "operationId": "AuthController_me_api_v1_auth_me_get",
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DoctorDetailRead"
    }
    }
    }
    }
    },
    "security": [
    {
    "APIKeyCookie": []
    }
    ]
    },
    "patch": {
    "tags": [
    "auth"
    ],
    "summary": "Authcontroller.Update Me",
    "operationId": "AuthController_update_me_api_v1_auth_me_patch",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DoctorUpdate"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DoctorDetailRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    },
    "security": [
    {
    "APIKeyCookie": []
    }
    ]
    }
    },
    "/api/v1/patients/": {
    "post": {
    "tags": [
    "patients"
    ],
    "summary": "Patientcontroller.Create Patient",
    "operationId": "PatientController_create_patient_api_v1_patients__post",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PatientCreate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PatientDetailRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "patients"
    ],
    "summary": "Patientcontroller.List Patients",
    "operationId": "PatientController_list_patients_api_v1_patients__get",
    "parameters": [
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_PatientRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/patients/{patient_id}": {
    "delete": {
    "tags": [
    "patients"
    ],
    "summary": "Patientcontroller.Delete Patient",
    "operationId": "PatientController_delete_patient_api_v1_patients__patient_id__delete",
    "parameters": [
    {
    "name": "patient_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Patient Id"
    }
    }
    ],
    "responses": {
    "204": {
    "description": "Successful response"
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "patients"
    ],
    "summary": "Patientcontroller.Get Patient",
    "operationId": "PatientController_get_patient_api_v1_patients__patient_id__get",
    "parameters": [
    {
    "name": "patient_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Patient Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PatientDetailRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "patch": {
    "tags": [
    "patients"
    ],
    "summary": "Patientcontroller.Patch Patient",
    "operationId": "PatientController_patch_patient_api_v1_patients__patient_id__patch",
    "parameters": [
    {
    "name": "patient_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Patient Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PatientUpdate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/PatientDetailRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
...
    "/api/v1/doctors/": {
    "post": {
    "tags": [
    "Doctors"
    ],
    "summary": "Doctorcontroller.Create Doctor",
    "operationId": "DoctorController_create_doctor_api_v1_doctors__post",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DoctorCreate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DoctorDetailRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Doctors"
    ],
    "summary": "Doctorcontroller.List Doctors",
    "operationId": "DoctorController_list_doctors_api_v1_doctors__get",
    "parameters": [
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_DoctorRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/doctors/{doctor_id}": {
    "delete": {
    "tags": [
    "Doctors"
    ],
    "summary": "Doctorcontroller.Delete Doctor",
    "operationId": "DoctorController_delete_doctor_api_v1_doctors__doctor_id__delete",
    "parameters": [
    {
    "name": "doctor_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Doctor Id"
    }
    }
    ],
    "responses": {
    "204": {
    "description": "Successful response"
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Doctors"
    ],
    "summary": "Doctorcontroller.Get Doctor",
    "operationId": "DoctorController_get_doctor_api_v1_doctors__doctor_id__get",
    "parameters": [
    {
    "name": "doctor_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Doctor Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DoctorDetailRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "patch": {
    "tags": [
    "Doctors"
    ],
    "summary": "Doctorcontroller.Update Doctor",
    "operationId": "DoctorController_update_doctor_api_v1_doctors__doctor_id__patch",
    "parameters": [
    {
    "name": "doctor_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Doctor Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DoctorUpdate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DoctorDetailRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
...
    "/api/v1/departments/": {
    "post": {
    "tags": [
    "Departments"
    ],
    "summary": "Departmentcontroller.Create Department",
    "operationId": "DepartmentController_create_department_api_v1_departments__post",
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DepartmentCreate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DepartmentRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Departments"
    ],
    "summary": "Departmentcontroller.List Departments",
    "operationId": "DepartmentController_list_departments_api_v1_departments__get",
    "parameters": [
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_DepartmentRead_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/departments/{department_id}": {
    "delete": {
    "tags": [
    "Departments"
    ],
    "summary": "Departmentcontroller.Delete Department",
    "operationId": "DepartmentController_delete_department_api_v1_departments__department_id__delete",
    "parameters": [
    {
    "name": "department_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "title": "Department Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "get": {
    "tags": [
    "Departments"
    ],
    "summary": "Departmentcontroller.Get Department",
    "operationId": "DepartmentController_get_department_api_v1_departments__department_id__get",
    "parameters": [
    {
    "name": "department_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "title": "Department Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DepartmentDetailRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    },
    "patch": {
    "tags": [
    "Departments"
    ],
    "summary": "Departmentcontroller.Patch Department",
    "operationId": "DepartmentController_patch_department_api_v1_departments__department_id__patch",
    "parameters": [
    {
    "name": "department_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "title": "Department Id"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DepartmentUpdate"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/DepartmentRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
...
    },
    "components": {
    "schemas": {
...
    "DepartmentCreate": {
    "properties": {
    "name": {
    "type": "string",
    "title": "Name"
    },
    "description": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Description"
    },
    "contact_phone": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Contact Phone"
    }
    },
    "type": "object",
    "required": [
    "name"
    ],
    "title": "DepartmentCreate"
    },
    "DepartmentDetailRead": {
    "properties": {
    "department_id": {
    "type": "integer",
    "title": "Department Id"
    },
    "name": {
    "type": "string",
    "title": "Name"
    },
    "description": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Description"
    },
    "contact_phone": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Contact Phone"
    },
    "doctors": {
    "items": {
    "$ref": "#/components/schemas/DoctorRead"
    },
    "type": "array",
    "title": "Doctors"
    }
    },
    "type": "object",
    "required": [
    "department_id",
    "name",
    "doctors"
    ],
    "title": "DepartmentDetailRead"
    },
    "DepartmentRead": {
    "properties": {
    "department_id": {
    "type": "integer",
    "title": "Department Id"
    },
    "name": {
    "type": "string",
    "title": "Name"
    },
    "description": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Description"
    },
    "contact_phone": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Contact Phone"
    }
    },
    "type": "object",
    "required": [
    "department_id",
    "name"
    ],
    "title": "DepartmentRead"
    },
    "DepartmentUpdate": {
    "properties": {
    "name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Name"
    },
    "description": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Description"
    },
    "contact_phone": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Contact Phone"
    }
    },
    "type": "object",
    "title": "DepartmentUpdate"
    },
    "DoctorCreate": {
    "properties": {
    "first_name": {
    "type": "string",
    "title": "First Name"
    },
    "last_name": {
    "type": "string",
    "title": "Last Name"
    },
    "middle_name": {
    "type": "string",
    "title": "Middle Name"
    },
    "email": {
    "type": "string",
    "title": "Email"
    },
    "phone_number": {
    "type": "string",
    "title": "Phone Number"
    },
    "birth_data": {
    "type": "string",
    "format": "date",
    "title": "Birth Data"
    },
    "gender": {
    "$ref": "#/components/schemas/Gender"
    },
    "specialization": {
    "type": "string",
    "title": "Specialization"
    },
    "department_id": {
    "type": "integer",
    "title": "Department Id"
    },
    "role_id": {
    "anyOf": [
    {
    "type": "integer"
    },
    {
    "type": "null"
    }
    ],
    "title": "Role Id"
    },
    "password": {
    "type": "string",
    "title": "Password"
    }
    },
    "type": "object",
    "required": [
    "first_name",
    "last_name",
    "middle_name",
    "email",
    "phone_number",
    "birth_data",
    "gender",
    "specialization",
    "department_id",
    "password"
    ],
    "title": "DoctorCreate"
    },
    "DoctorDetailRead": {
    "properties": {
    "doctor_id": {
    "type": "string",
    "format": "uuid",
    "title": "Doctor Id"
    },
    "first_name": {
    "type": "string",
    "title": "First Name"
    },
    "last_name": {
    "type": "string",
    "title": "Last Name"
    },
    "middle_name": {
    "type": "string",
    "title": "Middle Name"
    },
    "email": {
    "type": "string",
    "title": "Email"
    },
    "phone_number": {
    "type": "string",
    "title": "Phone Number"
    },
    "birth_data": {
    "type": "string",
    "format": "date",
    "title": "Birth Data"
    },
    "is_active": {
    "type": "boolean",
    "title": "Is Active"
    },
    "gender": {
    "$ref": "#/components/schemas/Gender"
    },
    "specialization": {
    "type": "string",
    "title": "Specialization"
    },
    "department": {
    "$ref": "#/components/schemas/DepartmentRead"
    },
    "role": {
    "$ref": "#/components/schemas/RoleRead"
    },
    "patients": {
    "items": {
    "$ref": "#/components/schemas/PatientRead"
    },
    "type": "array",
    "title": "Patients"
    }
    },
    "type": "object",
    "required": [
    "doctor_id",
    "first_name",
    "last_name",
    "middle_name",
    "email",
    "phone_number",
    "birth_data",
    "is_active",
    "gender",
    "specialization",
    "department",
    "role",
    "patients"
    ],
    "title": "DoctorDetailRead"
    },
...
    "DoctorUpdate": {
    "properties": {
    "first_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "First Name"
    },
    "last_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Last Name"
    },
    "middle_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Middle Name"
    },
    "email": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Email"
    },
    "phone_number": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Phone Number"
    },
    "birth_data": {
    "anyOf": [
    {
    "type": "string",
    "format": "date"
    },
    {
    "type": "null"
    }
    ],
    "title": "Birth Data"
    },
    "gender": {
    "anyOf": [
    {
    "$ref": "#/components/schemas/Gender"
    },
    {
    "type": "null"
    }
    ]
    },
    "specialization": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Specialization"
    },
    "department_id": {
    "anyOf": [
    {
    "type": "integer"
    },
    {
    "type": "null"
    }
    ],
    "title": "Department Id"
    },
    "role_id": {
    "anyOf": [
    {
    "type": "integer"
    },
    {
    "type": "null"
    }
    ],
    "title": "Role Id"
    },
    "password": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Password"
    }
    },
    "type": "object",
    "title": "DoctorUpdate"
    },
...
    "HTTPValidationError": {
    "properties": {
    "detail": {
    "items": {
    "$ref": "#/components/schemas/ValidationError"
    },
    "type": "array",
    "title": "Detail"
    }
    },
    "type": "object",
    "title": "HTTPValidationError"
    },
    "LoginCredentials": {
    "properties": {
    "username": {
    "type": "string",
    "title": "Username"
    },
    "password": {
    "type": "string",
    "title": "Password"
    }
    },
    "type": "object",
    "required": [
    "username",
    "password"
    ],
    "title": "LoginCredentials"
    },
...
    "Page_DepartmentRead_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/DepartmentRead"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "total",
    "page",
    "size"
    ],
    "title": "Page[DepartmentRead]"
    },
    "Page_DoctorRead_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/DoctorRead"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "total",
    "page",
    "size"
    ],
    "title": "Page[DoctorRead]"
    },
...
    "Page_PatientRead_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/PatientRead"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "total",
    "page",
    "size"
    ],
    "title": "Page[PatientRead]"
    },
...
    "PatientCreate": {
    "properties": {
    "first_name": {
    "type": "string",
    "title": "First Name"
    },
    "last_name": {
    "type": "string",
    "title": "Last Name"
    },
    "middle_name": {
    "type": "string",
    "title": "Middle Name"
    },
    "phone_number": {
    "type": "string",
    "title": "Phone Number"
    },
    "email": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Email"
    },
    "address": {
    "type": "string",
    "title": "Address"
    },
    "birth_data": {
    "type": "string",
    "format": "date",
    "title": "Birth Data"
    },
    "gender": {
    "$ref": "#/components/schemas/Gender"
    }
    },
    "type": "object",
    "required": [
    "first_name",
    "last_name",
    "middle_name",
    "phone_number",
    "address",
    "birth_data",
    "gender"
    ],
    "title": "PatientCreate"
    },
    "PatientDetailRead": {
    "properties": {
    "patient_id": {
    "type": "string",
    "format": "uuid",
    "title": "Patient Id"
    },
    "first_name": {
    "type": "string",
    "title": "First Name"
    },
    "last_name": {
    "type": "string",
    "title": "Last Name"
    },
    "middle_name": {
    "type": "string",
    "title": "Middle Name"
    },
    "phone_number": {
    "type": "string",
    "title": "Phone Number"
    },
    "email": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Email"
    },
    "birth_data": {
    "type": "string",
    "format": "date",
    "title": "Birth Data"
    },
    "is_active": {
    "type": "boolean",
    "title": "Is Active"
    },
    "address": {
    "type": "string",
    "title": "Address"
    },
    "gender": {
    "$ref": "#/components/schemas/Gender"
    },
    "medical_records": {
    "items": {
    "$ref": "#/components/schemas/MedRecordInPatient"
    },
    "type": "array",
    "title": "Medical Records",
    "default": []
    }
    },
    "type": "object",
    "required": [
    "patient_id",
    "first_name",
    "last_name",
    "middle_name",
    "phone_number",
    "email",
    "birth_data",
    "is_active",
    "address",
    "gender"
    ],
    "title": "PatientDetailRead"
    },
...
    "PatientUpdate": {
    "properties": {
    "first_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "First Name"
    },
    "last_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Last Name"
    },
    "middle_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Middle Name"
    },
    "phone_number": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Phone Number"
    },
    "email": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Email"
    },
    "address": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Address"
    },
    "birth_data": {
    "anyOf": [
    {
    "type": "string",
    "format": "date"
    },
    {
    "type": "null"
    }
    ],
    "title": "Birth Data"
    },
    "gender": {
    "anyOf": [
    {
    "$ref": "#/components/schemas/Gender"
    },
    {
    "type": "null"
    }
    ]
    }
    },
    "type": "object",
    "title": "PatientUpdate"
    },
...
    "ValidationError": {
    "properties": {
    "loc": {
    "items": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "integer"
    }
    ]
    },
    "type": "array",
    "title": "Location"
    },
    "msg": {
    "type": "string",
    "title": "Message"
    },
    "type": {
    "type": "string",
    "title": "Error Type"
    }
    },
    "type": "object",
    "required": [
    "loc",
    "msg",
    "type"
    ],
    "title": "ValidationError"
    }
    },
    "securitySchemes": {
    "APIKeyCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "med_voice_auth_token"
    }
    }
    }
    }